requirements.txt
altgraph==0.17.4
binarytree==6.5.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
git-filter-repo==2.38.0
graphviz==0.20.3
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pygame-ce==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyxel==2.0.7
requests==2.31.0
ruff==0.3.2
setuptools==69.1.1
setuptools-scm==8.1.0
six==1.16.0
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.1
